tools/ncs_reporter/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ncs-reporter"
version = "0.1.0"
description = "Standalone reporting CLI for NCS (Codex)"
dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
]
requires-python = ">=3.10"

[project.scripts]
ncs-reporter = "ncs_reporter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
ignore_missing_imports = true
strict = true
